+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All Notable changes to will be documented in this file
 
+## 1.0.2 - 2015-11-24
+### Changed
+- Added `src` folder with ES2015 code
+- Babel transformation to `lib/viewport-utility.js`
+
 ## 1.0.1 - 2015-11-24
-### Update
+### Changed
 - Changelog and Readme corrections
 
 ## 1.0.0 - 2015-11-24